Add migration methods for actor ID lookup
