+++ src/parser/core/Parser.tsx
@@ -10,6 +10,7 @@ import type {Actor as FflogsActor, Fight, Pet} from 'fflogs'
 import type {Event as LegacyEvent} from 'legacyEvent'
 import React from 'react'
 import {Report, Pull, Actor} from 'report'
+import {resolveActorId} from 'reportSources/legacyFflogs/base'
 import {Report as LegacyReport} from 'store/report'
 import toposort from 'toposort'
 import {extractErrorContext, isDefined, formatDuration} from 'utilities'
@@ -664,12 +665,38 @@ class Parser {
 
 	/**
 	 * Convert a unix epoch timestamp to a report-relative timestamp compatible
-	 * with logic running in the old analysis system.. Do ***not*** use this
+	 * with logic running in the old analysis system. Do ***not*** use this
 	 * outside the above scenario.
 	 * @deprecated
 	 */
 	epochToFflogs = (timestamp: number) =>
 		timestamp - this.pull.timestamp + this.eventTimeOffset
+
+	/**
+	 * Get an actor ID compatible with the new analysis system from the target of
+	 * a legacy fflogs event. This should ***only*** be utilised in old modules that
+	 * need to interop with the new system.
+	 * @deprecated
+	 */
+	getFflogsEventSourceActorId = (event: LegacyEvent) =>
+		resolveActorId({
+			id: event.sourceID,
+			instance: event.sourceInstance,
+			actor: event.source,
+		})
+
+	/**
+	 * Get an actor ID compatible with the new analysis system from the target of
+	 * a legacy fflogs event. This should ***only*** be utilised in old modules that
+	 * need to interop with the new system.
+	 * @deprecated
+	 */
+	getFflogsEventTargetActorId = (event: LegacyEvent) =>
+		resolveActorId({
+			id: event.targetID,
+			instance: event.targetInstance,
+			actor: event.target,
+		})
 
 	formatTimestamp(timestamp: number, secondPrecision?: number) {
 		return this.formatDuration(timestamp - this.fight.start_time, secondPrecision)
