+++ src/reportSources/legacyFflogs/eventAdapter/base.ts
@@ -1,19 +1,6 @@
 import {Event} from 'event'
-import {EventActor, FflogsEvent} from 'fflogs'
-import {Actor, Pull, Report} from 'report'
-
-export function resolveActorId(opts: {
-	id?: number,
-	instance?: number,
-	actor?: EventActor
-}): Actor['id'] {
-	const idNum = (opts.id ?? opts.actor?.id ?? -1)
-	const id = idNum === -1 ? 'unknown' : idNum.toString()
-	const instance = opts.instance ?? 1
-	return instance > 1
-		? `${id}:${instance}`
-		: id
-}
+import {FflogsEvent} from 'fflogs'
+import {Pull, Report} from 'report'
 
 export const PREPULL_OFFSETS = {
 	STATUS_ACTION: -3,
